requirements.txt
torch==2.6.0+cu124
torchvision==0.21.0+cu124
mmengine==0.10.7
mmcv==2.1.0
mmsegmentation==1.2.2
mamba-ssm==2.2.4
timm==1.0.15
einops==0.8.1
transformers==4.50.0
tensorboardX==2.6.2.2
numpy>=1.23
opencv-python>=4.8
Pillow==11.1.0
requests==2.32.3
ftfy
regex
tqdm
rich
pyyaml
